Split out tests and add more coverage
