+++ store/src/test/scala/geotrellis/store/index/zcurve/ZSpaceTimeLargeKeyIndexSpec.scala
@@ -23,15 +23,15 @@ import java.time.{ZoneOffset, ZonedDateTime}
 
 import org.scalatest._
 
-class ZSpaceTimeKeyIndexSpec extends FunSpec with Matchers {
+class ZSpaceTimeLargeKeyIndexSpec extends FunSpec with Matchers {
   val y2k = ZonedDateTime.of(2000, 1, 1, 0, 0, 0, 0, ZoneOffset.UTC)
   val upperBound = 8
   val keyBounds = KeyBounds(SpaceTimeKey(0, 0, 0L), SpaceTimeKey(100, 100, 100L))
 
-  describe("ZSpaceTimeKeyIndex test") {
+  describe("ZSpaceTimeLargeKeyIndex test") {
 
     it("indexes time") {
-      val zst = ZSpaceTimeKeyIndex.byYear(keyBounds)
+      val zst = ZSpaceTimeLargeKeyIndex.byYear(keyBounds)
 
       val keys =
         for (col <- 0 until upperBound;
@@ -46,7 +46,7 @@ class ZSpaceTimeKeyIndexSpec extends FunSpec with Matchers {
     }
 
     it("generates indexes you can check by hand 2x2x2") {
-      val zst = ZSpaceTimeKeyIndex.byMilliseconds(keyBounds, 1)
+      val zst = ZSpaceTimeLargeKeyIndex.byMilliseconds(keyBounds, 1)
       val idx = List[SpaceTimeKey](
         SpaceTimeKey(0, 0, y2k),
         SpaceTimeKey(1, 0, y2k),
@@ -65,7 +65,7 @@ class ZSpaceTimeKeyIndexSpec extends FunSpec with Matchers {
     }
 
     it("generates a Seq[(Long,Long)] from a keyRange: (SpaceTimeKey, SpaceTimeKey)") {
-      val zst = ZSpaceTimeKeyIndex.byMilliseconds(keyBounds, 1)
+      val zst = ZSpaceTimeLargeKeyIndex.byMilliseconds(keyBounds, 1)
 
       //all sub cubes in a 2x2x2
       var idx = zst.indexRanges((SpaceTimeKey(0, 0, y2k), SpaceTimeKey(1, 1, y2k.plus(1, MILLIS))))
@@ -112,7 +112,7 @@ class ZSpaceTimeKeyIndexSpec extends FunSpec with Matchers {
     }
 
     it("generates indexes by month") {
-      val zst = ZSpaceTimeKeyIndex.byMonth(keyBounds)
+      val zst = ZSpaceTimeLargeKeyIndex.byMonth(keyBounds)
 
       val keys =
         for (col <- 0 until upperBound;
@@ -124,6 +124,31 @@ class ZSpaceTimeKeyIndexSpec extends FunSpec with Matchers {
       keys.distinct.size should be(upperBound * upperBound * upperBound)
       keys.min should be(zst.toIndex(SpaceTimeKey(0, 0, y2k)))
       keys.max should be(zst.toIndex(SpaceTimeKey(upperBound - 1, upperBound - 1, y2k.plusMonths(upperBound - 1))))
+    }
+
+    // https://github.com/geotrellis/geotrellis-server/issues/248
+    it("should generate correct ranges outside of the unix time bounds") {
+      // layer dates range
+      // ["1960-01-01T12:00Z","1974-01-01T12:00Z"]
+
+      // unix time is in [1 January 1970; 1 January 2038) bounds
+      val minDate = ZonedDateTime.parse("1960-01-01T12:00Z")
+      val maxDate = ZonedDateTime.parse("1974-01-01T12:00Z")
+
+      minDate.toInstant.toEpochMilli shouldBe -315576000000L
+      maxDate.toInstant.toEpochMilli shouldBe 126273600000L
+
+      // -315576000000L
+      val minKey = SpaceTimeKey(4577, 5960, instant = minDate.toInstant.toEpochMilli)
+      // 126273600000L
+      val maxKey = SpaceTimeKey(4590, 5971, instant = maxDate.toInstant.toEpochMilli)
+      val kb = KeyBounds(minKey, maxKey)
+
+      // val temporalResolution: Long = 31536000000L
+      val index = ZSpaceTimeLargeKeyIndex.byYear(kb)
+      index.temporalResolution shouldBe 31536000000L
+
+      index.indexRanges(minKey -> maxKey).length shouldBe 401
     }
   }
 }
